# Reorganize queries and indexes for internal_transactions table