--- a/apps/explorer/lib/explorer/etherscan/logs.ex
+++ b/apps/explorer/lib/explorer/etherscan/logs.ex
@@ -5,8 +5,9 @@ defmodule Explorer.Etherscan.Logs do
 
   """
 
-  import Ecto.Query, only: [from: 2, where: 3, subquery: 1, order_by: 3]
+  import Ecto.Query, only: [from: 2, where: 3, subquery: 1, order_by: 3, union: 2]
 
+  alias Explorer.Chain
   alias Explorer.Chain.{Block, InternalTransaction, Log, Transaction}
   alias Explorer.Repo
 
@@ -78,25 +79,25 @@ defmodule Explorer.Etherscan.Logs do
 
     logs_query = where_topic_match(Log, prepared_filter)
 
+    query_to_address_hash_wrapped =
+      logs_query
+      |> internal_transaction_query(:to_address_hash, prepared_filter, address_hash)
+      |> Chain.wrapped_union_subquery()
+
+    query_from_address_hash_wrapped =
+      logs_query
+      |> internal_transaction_query(:from_address_hash, prepared_filter, address_hash)
+      |> Chain.wrapped_union_subquery()
+
+    query_created_contract_address_hash_wrapped =
+      logs_query
+      |> internal_transaction_query(:created_contract_address_hash, prepared_filter, address_hash)
+      |> Chain.wrapped_union_subquery()
+
     internal_transaction_log_query =
-      from(internal_transaction in InternalTransaction,
-        join: transaction in assoc(internal_transaction, :transaction),
-        join: log in ^logs_query,
-        on: log.transaction_hash == internal_transaction.transaction_hash,
-        where: internal_transaction.block_number >= ^prepared_filter.from_block,
-        where: internal_transaction.block_number <= ^prepared_filter.to_block,
-        where:
-          internal_transaction.to_address_hash == ^address_hash or
-            internal_transaction.from_address_hash == ^address_hash or
-            internal_transaction.created_contract_address_hash == ^address_hash,
-        select:
-          merge(map(log, ^@log_fields), %{
-            gas_price: transaction.gas_price,
-            gas_used: transaction.gas_used,
-            transaction_index: transaction.index,
-            block_number: transaction.block_number
-          })
-      )
+      query_to_address_hash_wrapped
+      |> union(^query_from_address_hash_wrapped)
+      |> union(^query_created_contract_address_hash_wrapped)
 
     all_transaction_logs_query =
       from(transaction in Transaction,
@@ -255,5 +256,23 @@ defmodule Explorer.Etherscan.Logs do
         data.index > ^log_index and data.block_number >= ^block_number and
           data.transaction_index >= ^transaction_index
     )
+  end
+
+  defp internal_transaction_query(logs_query, direction, prepared_filter, address_hash) do
+    from(internal_transaction in InternalTransaction,
+      join: transaction in assoc(internal_transaction, :transaction),
+      join: log in ^logs_query,
+      on: log.transaction_hash == internal_transaction.transaction_hash,
+      where: internal_transaction.block_number >= ^prepared_filter.from_block,
+      where: internal_transaction.block_number <= ^prepared_filter.to_block,
+      select:
+        merge(map(log, ^@log_fields), %{
+          gas_price: transaction.gas_price,
+          gas_used: transaction.gas_used,
+          transaction_index: transaction.index,
+          block_number: transaction.block_number
+        })
+    )
+    |> InternalTransaction.where_address_fields_match(address_hash, direction)
   end
 end